feat: 투자 성향 파악 게임 타입 오류 수정

See merge request s13-bigdata-dist-sub1/S13P21A301!84

+++ back/src/main/java/com/stock/survive/service/TendencyGameService.java
@@ -8,13 +8,13 @@ import com.stock.survive.dto.tendency.TendencyGameStateResponse;
 
 public interface TendencyGameService {
 
-    TendencyGameStateResponse start(Integer userId, TendencyGameStartRequest request);
+    TendencyGameStateResponse start(Long userId, TendencyGameStartRequest request);
 
-    TendencyGameStateResponse getState(Integer userId, Long sessionId);
+    TendencyGameStateResponse getState(Long userId, Long sessionId);
 
-    TendencyGameStateResponse placeOrder(Integer userId, Long sessionId, TendencyGameOrderRequest request);
+    TendencyGameStateResponse placeOrder(Long userId, Long sessionId, TendencyGameOrderRequest request);
 
-    TendencyGameStateResponse proceedNextWeek(Integer userId, Long sessionId);
+    TendencyGameStateResponse proceedNextWeek(Long userId, Long sessionId);
 
-    TendencyGameResultResponse finish(Integer userId, TendencyGameFinishRequest request);
+    TendencyGameResultResponse finish(Long userId, TendencyGameFinishRequest request);
 }
